+++ src/CrossValidation/MonteCarlo.php
@@ -3,9 +3,13 @@
 namespace Rubix\ML\CrossValidation;
 
 use Rubix\ML\Learner;
+use Rubix\ML\Parallel;
 use Rubix\ML\Estimator;
+use Rubix\ML\Backends\Serial;
 use Rubix\ML\Datasets\Labeled;
+use Rubix\ML\Datasets\Dataset;
 use Rubix\ML\Other\Helpers\Stats;
+use Rubix\ML\Other\Traits\Multiprocessing;
 use Rubix\ML\CrossValidation\Metrics\Metric;
 use Rubix\ML\Other\Specifications\EstimatorIsCompatibleWithMetric;
 use InvalidArgumentException;
@@ -21,8 +25,10 @@ use InvalidArgumentException;
  * @package     Rubix/ML
  * @author      Andrew DalPino
  */
-class MonteCarlo implements Validator
+class MonteCarlo implements Validator, Parallel
 {
+    use Multiprocessing;
+
     /**
      * The number of simulations to run i.e the number of tests to average.
      *
@@ -65,6 +71,7 @@ class MonteCarlo implements Validator
         $this->simulations = $simulations;
         $this->ratio = $ratio;
         $this->stratify = $stratify;
+        $this->backend = new Serial();
     }
 
     /**
@@ -89,13 +96,32 @@ class MonteCarlo implements Validator
                 ? $dataset->stratifiedSplit($this->ratio)
                 : $dataset->split($this->ratio);
     
-            $estimator->train($training);
-    
-            $predictions = $estimator->predict($testing);
-    
-            $scores[] = $metric->score($predictions, $testing->labels());
+            $this->backend->enqueue(
+                [self::class, 'score'],
+                [$estimator, $training, $testing, $metric]
+            );
         }
+    
+        $scores = $this->backend->process();
 
         return Stats::mean($scores);
+    }
+
+    /**
+     * Score an estimator on one of n simulations.
+     *
+     * @param \Rubix\ML\Learner $estimator
+     * @param \Rubix\ML\Datasets\Dataset $training
+     * @param \Rubix\ML\Datasets\Labeled $testing
+     * @param \Rubix\ML\CrossValidation\Metrics\Metric $metric
+     * @return float
+     */
+    public static function score(Learner $estimator, Dataset $training, Labeled $testing, Metric $metric) : float
+    {
+        $estimator->train($training);
+
+        $predictions = $estimator->predict($testing);
+
+        return $metric->score($predictions, $testing->labels());
     }
 }